+++ src/build-apps.js
@@ -1,20 +1,26 @@
-const path = require('path');
+import path from 'path';
 
 let chalk;
-const spawn = require('cross-spawn');
-const fs = require('fs-extra');
-const readdirp = require('readdirp');
+import spawn from 'cross-spawn';
+import fs from 'fs-extra';
+import {readdirpPromise} from 'readdirp';
+import * as url from 'url';
 
 const env = {
 	ILIB_BASE_PATH: '/framework/ilib',
 	ILIB_ASSET_CREATE: 'false',
 	SIMPLE_CSS_IDENT: 'true',
-	BROWSERSLIST: 'Chrome 79'
+	BROWSERSLIST: 'Chrome 132'
 };
 
 function findViews (base) {
-	return readdirp.promise(path.join('tests', base, 'apps'), {fileFilter: '*-View.js'});
+	return readdirpPromise(path.join('tests', base, 'apps'), {
+		fileFilter: (entry) => entry.basename.endsWith('-View.js')
+	});
 }
+
+// eslint-disable-next-line no-shadow
+const __dirname = path.dirname(url.fileURLToPath(import.meta.url));
 
 function buildApps (base) {
 	if (process.argv.includes('--skip-build')) return;
@@ -140,5 +146,6 @@ function epack ({file, opts}) {
 	}
 }
 
-module.exports = buildApps;
-if (require.main === module) buildApps();
+export default buildApps;
+const modulePath = url.fileURLToPath(import.meta.url);
+if (process.argv[1] === modulePath) buildApps();