+++ ui/wdio.tv.conf.js
@@ -1,16 +1,18 @@
-const ipAddress = require('../utils/ipAddress.js');
-const {config} = require('./wdio.conf.js');
+import {ipAddress} from '../utils/ipAddress.js';
+import {config as uiConfig} from './wdio.conf.js';
 
-exports.config = Object.assign(
+const config = Object.assign(
 	{},
-	config,
+	uiConfig,
 	{
 		capabilities: [{
-			maxInstances: 1,
-
+			// maxInstances can get overwritten per capability. So if you have an in-house Selenium
+			// grid with only 5 firefox instances available you can make sure that not more than
+			// 5 instances get started at a time.
+			'wdio:maxInstances': 1,
+			//
 			browserName: 'chrome',
 			'goog:chromeOptions': {
-				w3c: false,
 				debuggerAddress: `${process.env.TV_IP}:9998`
 			}
 		}],
@@ -18,11 +20,13 @@ exports.config = Object.assign(
 		baseUrl: `http://${ipAddress()}:4567`,
 
 		before: function () {
-			if (config.before) {
-				config.before();
+			if (uiConfig.before) {
+				uiConfig.before();
 			}
 
 			browser._options = {remote: true};
 		}
 	}
 );
+
+export const uiTVConfig = {config};