+++ utils/Page.js
@@ -1,8 +1,6 @@
-'use strict';
+import queryString from 'query-string';
 
-const queryString = require('query-string');
-
-class Page {
+export class Page {
 	constructor () {
 		this.title = 'Untitled Test';
 		this._url = '';
@@ -14,19 +12,15 @@ class Page {
 
 	async open (appPath, urlExtra = '?locale=en-US') {
 		await browser.execute(function () {
-			document.body.innerHTML = '';
+			document.body.textContent = '';
 		});
 
 		this._url = `/${appPath}/${urlExtra}`;
-		// Can't resize browser window when connected to remote debugger!
-		if (!browser._options || !browser._options.remote) {
-			await browser.setWindowSize(1920, 1080);
-		}
 
 		await browser.url(this.url);
 
 		const body = await $('body');
-		await body.waitForDisplayed({timeout: 5000});
+		await body.waitForDisplayed({timeout: 10000});
 
 		await this.delay(200);
 	}
@@ -170,7 +164,3 @@ class Page {
 		);
 	}
 }
-
-module.exports = {
-	Page
-};